+++ android/src/main/kotlin/com/marvinvogl/n_gamepad/MotionListener.kt
@@ -10,17 +10,20 @@ class MotionListener(
     private val connection: Connection,
 ) : OnGenericMotionListener {
     companion object {
-        val buffer = ControlBuffer(26)
+        val buffer = ControlBuffer(28, 2, 0b1000)
     }
 
     override fun onGenericMotion(v: View?, event: MotionEvent?): Boolean {
         if (event != null) {
-            if (event.source and InputDevice.SOURCE_JOYSTICK == InputDevice.SOURCE_JOYSTICK) {
+            if (event.isFromSource(InputDevice.SOURCE_JOYSTICK)) {
                 gamepad.dpad.onEvent(event)
-                gamepad.joystickLeft.onEvent(event)
-                gamepad.joystickRight.onEvent(event)
+
+                connection.send(KeyListener.buffer)
+
                 gamepad.triggerLeft.onEvent(event)
                 gamepad.triggerRight.onEvent(event)
+                gamepad.joystickLeft.onEvent(event)
+                gamepad.joystickRight.onEvent(event)
 
                 return connection.send(buffer)
             }